Save for emergency

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: cmateos- <cmateos-@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/28 19:23:13 by cmateos-          #+#    #+#             */
-/*   Updated: 2023/05/21 21:45:02 by cmateos-         ###   ########.fr       */
+/*   Updated: 2023/05/22 18:06:55 by cmateos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef FT_PRINTF_H
@@ -22,7 +22,7 @@ int	ft_putchr_pf(char c);
 int	ft_putstr_pf(char *s);
 int	ft_putnbr_pf(long nbr, int count);
 int	ft_putptr_pf(unsigned long nbr, char *hex, int count);
-int	ft_puthex(unsigned long nbr, char *hex, int count);
+int	ft_puthex(unsigned int nbr, char *hex, int count);
 int	ft_putun(unsigned long, int count);
 
 #endif
